working on implementing more accurate behaviours for shooting

+++ app/src/main/java/bham/student/txm683/heartbreaker/entities/weapons/BasicWeapon.java
@@ -24,7 +24,7 @@ public class BasicWeapon extends Weapon {
         this.bulletLife = 25;
     }
 
-    public BasicWeapon(String owner, int cooldownTicks, int damage, int radiusModifier){
+    public BasicWeapon(String owner, int cooldownTicks, int damage, float radiusModifier){
         this(owner);
         setAfterFiringCooldownInTicks(cooldownTicks);
         this.damage = damage;
